app/backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.0
pillow==11.0.0
httpx==0.28.1
pydantic-settings==2.7.0
python-multipart==0.0.21
pytest==8.3.4
pytest-asyncio==0.25.0